separated commit function into separate file
